+++ stackit/internal/services/iaasalpha/routingtable/route/resource_test.go
@@ -1,4 +1,4 @@
-package routingtable_route
+package route
 
 import (
 	"context"
@@ -6,13 +6,14 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaasalpha/routingtable/shared"
+
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/stackitcloud/stackit-sdk-go/core/utils"
 	"github.com/stackitcloud/stackit-sdk-go/services/iaasalpha"
-	"github.com/stackitcloud/terraform-provider-stackit/stackit/internal/services/iaasalpha/shared"
 )
 
 const (
